# openapi.yaml
openapi: 3.1.0
info:
  title: VulnYoga API
  description: |
    # VulnYoga - Vulnerable-by-Design Yoga Store API
    
    ⚠️ **WARNING**: This API is intentionally vulnerable for security testing purposes.
    Do NOT expose this API to the internet or production environments.
    For isolated lab use only.
    
    ## OWASP API Security Top 10 (2023) Vulnerabilities
    
    This API intentionally includes vulnerabilities from the OWASP API Security Top 10:
    
    - **API1**: Broken Object Level Authorization (BOLA)
    - **API2**: Broken Authentication
    - **API3**: Broken Object Property Level Authorization (BOPLA)
    - **API4**: Unrestricted Resource Consumption
    - **API5**: Broken Function Level Authorization
    - **API6**: Unrestricted Access to Sensitive Business Flows
    - **API7**: Server-Side Request Forgery (SSRF)
    - **API8**: Security Misconfiguration
    - **API9**: Improper Inventory Management
    - **API10**: Unsafe Consumption of APIs
    
    Each endpoint is annotated with the vulnerabilities it demonstrates.
  version: 1.0.0
  contact:
    name: VulnYoga Team
    email: security@vulnyoga.local
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
paths:
  /api/v1/auth/register:
    post:
      summary: Register a new user
      x-owasp:
        categories: ["API3"]
        vulnerableWhenEnv: "VULN_API3_BOPLA=true"
        description: "Vulnerable to mass assignment of role field"
      tags: ["Authentication"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/login:
    post:
      summary: Login user
      x-owasp:
        categories: ["API2"]
        vulnerableWhenEnv: "VULN_API2_BROKEN_AUTH=true"
        description: "Weak JWT, accepts expired tokens, accepts token in query params"
      tags: ["Authentication"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/reset:
    post:
      summary: Request password reset
      x-owasp:
        categories: ["API2"]
        vulnerableWhenEnv: "VULN_API2_BROKEN_AUTH=true"
        description: "No proof of ownership required"
      tags: ["Authentication"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Reset email sent (if email exists)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /api/v1/auth/reset/consume:
    get:
      summary: Consume password reset token
      x-owasp:
        categories: ["API2"]
        vulnerableWhenEnv: "VULN_API2_BROKEN_AUTH=true"
        description: "Weak token verification, no expiration check"
      tags: ["Authentication"]
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsumeResetRequest'
      responses:
        '200':
          description: Password updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /api/v1/users/{id}:
    get:
      summary: Get user by ID
      x-owasp:
        categories: ["API1"]
        vulnerableWhenEnv: "VULN_API1_BOLA=true"
        description: "No ownership check - any authenticated user can access any user"
      tags: ["Users"]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: token
          in: query
          schema:
            type: string
          description: "Vulnerable: Token can be passed in query parameter"
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      summary: Update user
      x-owasp:
        categories: ["API1", "API3"]
        vulnerableWhenEnv: "VULN_API1_BOLA=true,VULN_API3_BOPLA=true"
        description: "No ownership check + mass assignment of sensitive fields"
      tags: ["Users"]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /api/v1/items:
    get:
      summary: Get all items
      tags: ["Items"]
      responses:
        '200':
          description: List of items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Item'

    post:
      summary: Create new item
      x-owasp:
        categories: ["API3", "API5"]
        vulnerableWhenEnv: "VULN_API3_BOPLA=true,VULN_API5_FUNC_AUTH=true"
        description: "Mass assignment + client-controlled role authorization"
      tags: ["Items"]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateItemRequest'
      responses:
        '201':
          description: Item created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
  /api/v1/items/search:
    get:
      summary: Search items
      x-owasp:
        categories: ["API4"]
        vulnerableWhenEnv: "VULN_API4_RESOURCE=true"
        description: "No limits on pageSize, CPU-intensive search"
      tags: ["Items"]
      parameters:
        - name: q
          in: query
          schema:
            type: string
          description: Search query
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: Page number
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 10
          description: "Vulnerable: No upper limit enforced"
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
  /api/v1/items/{id}:
    get:
      summary: Get item by ID
      tags: ["Items"]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Item details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'

    patch:
      summary: Update item
      x-owasp:
        categories: ["API3"]
        vulnerableWhenEnv: "VULN_API3_BOPLA=true"
        description: "Mass assignment of sensitive fields (costPrice, supplierEmail)"
      tags: ["Items"]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateItemRequest'
      responses:
        '200':
          description: Item updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'

    delete:
      summary: Delete item
      x-owasp:
        categories: ["API5"]
        vulnerableWhenEnv: "VULN_API5_FUNC_AUTH=true"
        description: "Client-controlled role authorization via X-Role header"
      tags: ["Items"]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: X-Role
          in: header
          schema:
            type: string
          description: "Vulnerable: Client can control authorization"
      responses:
        '200':
          description: Item deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /api/v1/orders:
    get:
      summary: Get user's orders
      tags: ["Orders"]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of user's orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'

    post:
      summary: Create new order
      tags: ["Orders"]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
  /api/v1/orders/{orderId}:
    get:
      summary: Get order by ID
      x-owasp:
        categories: ["API1"]
        vulnerableWhenEnv: "VULN_API1_BOLA=true"
        description: "No ownership check - any authenticated user can access any order"
      tags: ["Orders"]
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'

  # Cart Endpoints
  /api/v1/cart:
    get:
      summary: Get user's cart
      tags: ["Cart"]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User's cart contents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'

    post:
      summary: Add item to cart
      tags: ["Cart"]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddToCartRequest'
      responses:
        '200':
          description: Item added to cart successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'

    delete:
      summary: Remove item from cart
      tags: ["Cart"]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RemoveFromCartRequest'
      responses:
        '200':
          description: Item removed from cart successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
  /api/v1/cart/clear:
    delete:
      summary: Clear entire cart
      tags: ["Cart"]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Cart cleared successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
  /api/v1/cart-view/{userId}:
    get:
      summary: View cart by user ID (BOLA Vulnerability)
      x-owasp:
        categories: ["API1"]
        vulnerableWhenEnv: "VULN_API1_BOLA=true"
        description: "Broken Object Level Authorization - unauthenticated access to any user's cart"
      tags: ["Cart", "Vulnerabilities"]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
          description: ID of user whose cart to view
      responses:
        '200':
          description: Cart contents (vulnerable to BOLA)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        '404':
          description: Cart not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/checkout/pay:
    post:
      summary: Process payment
      x-owasp:
        categories: ["API6", "API10"]
        vulnerableWhenEnv: "VULN_API6_BUSINESS_FLOW=true,VULN_API10_UNSAFE_CONSUMP=true"
        description: "Out-of-order operations + trusts client-controlled payment status"
      tags: ["Checkout"]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckoutRequest'
      responses:
        '200':
          description: Payment processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
  /api/v1/checkout/ship:
    post:
      summary: Ship order
      x-owasp:
        categories: ["API6"]
        vulnerableWhenEnv: "VULN_API6_BUSINESS_FLOW=true"
        description: "Allows shipping before payment"
      tags: ["Checkout"]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                orderId:
                  type: integer
      responses:
        '200':
          description: Order shipped successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
  /api/v1/image/proxy:
    get:
      summary: Proxy image from URL
      x-owasp:
        categories: ["API7"]
        vulnerableWhenEnv: "VULN_API7_SSRF=true"
        description: "No URL validation, allows internal network access"
      tags: ["Utilities"]
      parameters:
        - name: url
          in: query
          required: true
          schema:
            type: string
          description: "Vulnerable: No allowlist, allows file://, internal IPs"
      responses:
        '200':
          description: Proxied image or content
          content:
            image/*:
              schema:
                type: string
                format: binary
        '400':
          description: Invalid URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/file/read:
    get:
      summary: Read local file
      x-owasp:
        categories: ["LFI"]
        vulnerableWhenEnv: "VULN_LFI=true"
        description: "No path validation, allows directory traversal attacks"
      tags: ["Utilities"]
      parameters:
        - name: file
          in: query
          required: true
          schema:
            type: string
          description: "Vulnerable: No path sanitization, allows ../../etc/passwd"
      responses:
        '200':
          description: File content
          content:
            text/plain:
              schema:
                type: string
            application/json:
              schema:
                type: string
            text/html:
              schema:
                type: string
        '400':
          description: Invalid file parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: File not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied (when secure mode)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/users/search:
    get:
      summary: Search users by username
      x-owasp:
        categories: ["SQLi"]
        vulnerableWhenEnv: "VULN_SQLI=true"
        description: "SQL Injection via unsanitized username parameter in raw SQL query"
      tags: ["Utilities"]
      parameters:
        - name: username
          in: query
          required: true
          schema:
            type: string
          description: "Vulnerable: Direct string concatenation in raw SQL query"
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
        '400':
          description: Missing username parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/redirect:
    get:
      summary: Redirect to URL
      x-owasp:
        categories: ["OpenRedirect"]
        vulnerableWhenEnv: "VULN_OPEN_REDIRECT=true"
        description: "Open redirect via unvalidated url parameter"
      tags: ["Utilities"]
      parameters:
        - name: url
          in: query
          required: true
          schema:
            type: string
          description: "Vulnerable: No URL validation, allows redirect to any external domain"
      responses:
        '302':
          description: Redirect to the specified URL
        '400':
          description: Missing or invalid URL parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/reports/sales:
    get:
      summary: Generate sales report
      x-owasp:
        categories: ["SQLi"]
        vulnerableWhenEnv: "VULN_SQLI=true"
        description: "SQL Injection via sortBy parameter in ORDER BY clause"
      tags: ["Reports"]
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
          description: Report start date
        - name: endDate
          in: query
          schema:
            type: string
          description: Report end date
        - name: region
          in: query
          schema:
            type: string
          description: Filter by region
        - name: category
          in: query
          schema:
            type: string
          description: Filter by category
        - name: sortBy
          in: query
          required: true
          schema:
            type: string
          description: "Vulnerable: Injected directly into ORDER BY clause"
      responses:
        '200':
          description: Sales report data
          content:
            application/json:
              schema:
                type: object
                properties:
                  report:
                    type: array
                    items:
                      $ref: '#/components/schemas/Item'
                  filters:
                    type: object
  /api/v1/templates/render:
    get:
      summary: Render a UI template
      x-owasp:
        categories: ["LFI"]
        vulnerableWhenEnv: "VULN_LFI=true"
        description: "LFI via theme parameter used as file path"
      tags: ["Utilities"]
      parameters:
        - name: theme
          in: query
          required: true
          schema:
            type: string
          description: "Vulnerable: Used directly as file path without sanitization"
        - name: lang
          in: query
          schema:
            type: string
          description: Language code
        - name: format
          in: query
          schema:
            type: string
          description: Output format
        - name: version
          in: query
          schema:
            type: string
          description: Template version
        - name: widget
          in: query
          schema:
            type: string
          description: Widget identifier
  /api/v1/auth/sso:
    get:
      summary: SSO callback handler
      x-owasp:
        categories: ["OpenRedirect"]
        vulnerableWhenEnv: "VULN_OPEN_REDIRECT=true"
        description: "Open redirect via landingPage parameter"
      tags: ["Authentication"]
      parameters:
        - name: provider
          in: query
          schema:
            type: string
          description: SSO provider name
        - name: tenantId
          in: query
          schema:
            type: string
          description: Tenant identifier
        - name: sessionId
          in: query
          schema:
            type: string
          description: Session identifier
        - name: landingPage
          in: query
          required: true
          schema:
            type: string
          description: "Vulnerable: Redirect target with no validation"
  /api/v1/export/csv:
    get:
      summary: Export items as CSV
      x-owasp:
        categories: ["API8"]
        vulnerableWhenEnv: "VULN_API8_MISCONFIG=true"
        description: "No authentication required for sensitive operation"
      tags: ["Export"]
      responses:
        '200':
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
        '401':
          description: Authentication required (when secure)
  /api/v1/admin/users:
    get:
      summary: Get all users (admin only)
      x-owasp:
        categories: ["API5"]
        vulnerableWhenEnv: "VULN_API5_FUNC_AUTH=true"
        description: "Any authenticated user can access admin endpoint"
      tags: ["Admin"]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of all users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
...
  /api/v0/items/bulk:
    get:
      summary: Legacy endpoint - Get items with optional supplier data
      x-owasp:
        categories: ["API9"]
        vulnerableWhenEnv: "VULN_API9_INVENTORY=true"
        description: "Legacy endpoint that leaks sensitive supplier data"
      tags: ["Legacy"]
      parameters:
        - name: include
          in: query
          schema:
            type: string
            enum: [supplier]
          description: "Vulnerable: Leaks costPrice and supplierEmail when include=supplier"
      responses:
        '200':
          description: List of items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Item'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Vulnerable: 
        - Weak JWT secret
        - Accepts expired tokens
        - Accepts token in query parameter (?token=...)
  schemas:
...
    CreateUserRequest:
      type: object
      required:
        - email
        - password
        - name
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
        address:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [CUSTOMER, STAFF, ADMIN]
          description: "Vulnerable: Mass assignment allowed"
...
    AuthResponse:
      type: object
      properties:
        message:
          type: string
        user:
          $ref: '#/components/schemas/User'
        token:
          type: string
...